% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2021-11-12_A1_JF_csp_cost_functions.R
\name{add_las_attributes}
\alias{add_las_attributes}
\title{Add all las_attributes from las@data to the header of a las element}
\usage{
add_las_attributes(las)
}
\arguments{
\item{las}{an element of lidR::LAS class}
}
\value{

}
\description{
The helper function adds all headings from las@data which are nor part of
lidR:::LASATTRIBUTES to the las header using lidR::add_lasattribute. Only
attributes that are included in the header got saved when using
lidR::writeLAS, this is a convenient way to add them.
}
\note{
%% ~~further notes~~
}
\examples{

LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
las <- lidR::readLAS(LASfile, select = "xyz", filter = "-inside 481250 3812980 481300 3813030")

las@data$noise <- runif(nrow(las@data))
las@data$noiseZ <- las@data$var1 * las@data$Z

las <- add_las_attributes(las)

}
\author{
Dr. Julian Frey <julian.frey@iww.uni-freiburg.de>
}
