% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2021-11-12_A1_JF_csp_cost_functions.R
\name{add_voxel_coordinates}
\alias{add_voxel_coordinates}
\title{Add voxel coordinates to a las file}
\usage{
add_voxel_coordinates(las, res)
}
\arguments{
\item{las}{an element of lidR::LAS class}

\item{res}{voxel ressolution in [m]}
}
\value{

}
\description{
Adds the collums x_vox, y_vox and z_vox in the given ressolution to the las
element. This is convinient if informations have been derived in voxel space
and these should be attached to the original points.
}
\details{
Voxel coordinates derived with this function are identical as those derived
by lidR::voxelize.
}
\note{
%% ~~further notes~~
}
\examples{

LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
las <- lidR::readLAS(LASfile, select = "xyz", filter = "-inside 481250 3812980 481300 3813030")

las <- add_voxel_coordinates(las,res = 1)

lidR::plot(las, color = 'z_vox')

}
\author{
Dr. Julian Frey <julian.frey@iww.uni-freiburg.de>
}
