% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigen_decomposition}
\alias{eigen_decomposition}
\title{Fast Eigenvalues decomposition for k nearest neighbors using a C++ function}
\usage{
eigen_decomposition(las, k, ncpu = 1L)
}
\arguments{
\item{las}{LAS element}

\item{k}{k nearest neighbors}

\item{ncpu}{number of cpu cores to use}
}
\value{
The function returns for every point the 3 eigenvalues and the
third element of the third eigenvector. These values are needed to compute
planarity, linerity, verticality etc. in the add_geometry function
}
\description{
C++ helper function to compute eigenvalues for geometric feature
calculation.
}
\seealso{
\link{add_geometry}
}
\author{
Julian Frey <julian.frey@iww.uni-freiburg.de>
}
