% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_unlist}
\alias{fast_unlist}
\title{helper function to unlist IDs generated by dbscan::frNN}
\usage{
fast_unlist(list, l)
}
\arguments{
\item{list}{a list element created by dbscan::frNN}

\item{l}{the expected length of the result}
}
\value{

}
\description{
helper function to unlist IDs generated by dbscan::frNN
}
