% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_unlist_dist}
\alias{fast_unlist_dist}
\title{helper function to unlist distances computed by dbscan::frNN}
\usage{
fast_unlist_dist(list, l)
}
\arguments{
\item{list}{a list element created by dbscan::frNN}

\item{l}{the expected length of the result}
}
\value{
Returns a vector with the values in the list.
}
\description{
extracts the distances from a nested list created by dbscan::frNN
}
\author{
Julian Frey <julian.frey@iww.uni-freiburg.de>
}
