% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_inventory.R
\name{fi_convex_hull_area}
\alias{fi_convex_hull_area}
\title{Compute convex hull area}
\usage{
fi_convex_hull_area(x, y)
}
\arguments{
\item{x}{Numeric vector of x coordinates.}

\item{y}{Numeric vector of y coordinates.}
}
\value{
Numeric area of the convex hull polygon.
}
\description{
Compute convex hull area
}
\keyword{internal}
