% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_inventory.R
\name{fi_fit_circle}
\alias{fi_fit_circle}
\title{Fit one circle from a tree slice}
\usage{
fi_fit_circle(slice, use_stem_segmentation = FALSE)
}
\arguments{
\item{slice}{A data.table with `X`, `Y`, `Planarity`, and `Verticality`.}

\item{use_stem_segmentation}{Logical. If `TRUE`, uses all slice points.}
}
\value{
Numeric vector `c(X, Y, r)`.
}
\description{
Fit one circle from a tree slice
}
\keyword{internal}
