% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_inventory.R
\name{fi_na2true}
\alias{fi_na2true}
\title{Normalize invalid checks to TRUE}
\usage{
fi_na2true(x)
}
\arguments{
\item{x}{A logical or numeric vector.}
}
\value{
A logical vector.
}
\description{
Treats `NA` and infinite values as `TRUE` to simplify guard conditions.
}
\keyword{internal}
