% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_inventory.R
\name{fi_process_trunk_slices}
\alias{fi_process_trunk_slices}
\title{Process trunk slices for stem quality}
\usage{
fi_process_trunk_slices(las)
}
\arguments{
\item{las}{A `lidR` LAS object for one tree.}
}
\value{
A data.frame with fitted circle parameters per vertical slice.
}
\description{
Process trunk slices for stem quality
}
\keyword{internal}
