% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_inventory.R
\name{fi_stem_quality_inventory}
\alias{fi_stem_quality_inventory}
\title{Build a stem quality inventory}
\usage{
fi_stem_quality_inventory(las, tree_id_col = "TreeID")
}
\arguments{
\item{las}{A `lidR` LAS object containing segmented trees.}

\item{tree_id_col}{Character column name containing tree ids.}
}
\value{
A data.frame with tree id, curvature, and taper.
}
\description{
Build a stem quality inventory
}
\keyword{internal}
