% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_inventory.R
\name{fi_zdif}
\alias{fi_zdif}
\title{Compute height span}
\usage{
fi_zdif(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric span `max(x) - min(x)`.
}
\description{
Compute height span
}
\keyword{internal}
