% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2021-11-12_A1_JF_csp_cost_functions.R
\name{find_base_coordinates_raster}
\alias{find_base_coordinates_raster}
\title{Find stem base position using a density raster approach}
\usage{
find_base_coordinates_raster(
  las,
  res = 0.1,
  zmin = 0.5,
  zmax = 2,
  q = 0.975,
  eps = 0.2
)
}
\arguments{
\item{las}{an element of lidR::LAS class}

\item{res}{raster resolution}

\item{zmin}{lower search boundary}

\item{zmax}{upper search boundary}

\item{quantile}{raster density quantile to assign a tree region}

\item{merge_radius}{search radius to merge base points}
}
\value{

}
\description{
Find stem base position using a density raster approach
}
\author{
Dr. Julian Frey <julian.frey@iww.uni-freiburg.de>
}
