% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_colors.R
\name{get_pal}
\alias{get_pal}
\title{Return a palette-generating function}
\usage{
get_pal(name = "rainbow")
}
\arguments{
\item{name}{Character scalar. Name of the palette to use. Supported names:
"sky", "sea", "cozy", "fairy", "winter", "rainbow", "pastel", "candy",
"boring". Default: "rainbow".}
}
\value{
A function with signature \code{function(n)} that returns \code{n}
  hex colour strings.
}
\description{
get_pal returns a function that produces a character vector of hex colours
given a required length. It exposes a set of named palettes used by
color_ids().
}
\examples{
get_pal("sky")

}
\author{
Zoe Schindler
}
