% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farthest_point_sampling.R
\name{p_dist}
\alias{p_dist}
\title{Point distance function}
\usage{
p_dist(p1, p2)
}
\arguments{
\item{p1}{point 1}

\item{p2}{point 2}
}
\value{
the distance between the two points
}
\description{
calculates euclidean distances for n dimensions
}
\examples{
p_dist(c(0,0), c(3,4))
}
