% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2024-06-10_A1_JF_farthest_point_sampling.R
\name{p_dist2}
\alias{p_dist2}
\title{Point distance function}
\usage{
p_dist2(mat, p)
}
\arguments{
\item{mat}{point 1}

\item{p}{point 2}
}
\value{
the distance between the two points
}
\description{
calculates euclidean distances for n dimensions between a matrix of points and a single point
}
\examples{
dist(c(0,0), c(3,4))
}
