% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_mat_dist}
\alias{p_mat_dist}
\title{Point distance function}
\usage{
p_mat_dist(mat, p, nthreads = 0L)
}
\arguments{
\item{mat}{matrix with points as rows}

\item{p}{point to calculate distances}

\item{nthreads}{number of threads to use. If 0 or negative, the maximum number of threads available will be used.}
}
\value{
the distances between every row of mat and p
}
\description{
calculates euclidean distances for n dimensions between a matrix of points and a single point
}
\examples{
p_mat_dist(as.matrix(cbind(runif(100),runif(100))), c(3,4))
}
