% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_inventory.R
\name{point_center_angle}
\alias{point_center_angle}
\title{Returns the angle between the center of the circle and a point in degrees}
\usage{
point_center_angle(point, circle)
}
\arguments{
\item{point}{numeric vector of length 2 c(X,Y)}

\item{circle}{numeric vector of length 3 c(center_X, center_Y, radius)}
}
\value{
numeric angle in degrees
}
\description{
Returns the angle between the center of the circle and a point in degrees
}
