% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_inventory.R
\name{point_circle_distance}
\alias{point_circle_distance}
\title{Helper function to compute distances from a point to the circle}
\usage{
point_circle_distance(point, circle)
}
\arguments{
\item{point}{numeric vector of length 2 c(X,Y)}

\item{circle}{numeric vector of length 3 c(center_X, center_Y, radius)}
}
\value{
numeric distance from the point to the circle
}
\description{
Helper function to compute distances from a point to the circle
}
