% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2025-01-10_A1_JF_forest_inventory.R
\name{suppress_cat}
\alias{suppress_cat}
\title{Suppress only the cat() output}
\usage{
suppress_cat(f, ...)
}
\arguments{
\item{f}{function to be called}
}
\value{
the return value of the function
}
\description{
Suppress only the cat() output
}
