% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2021-11-12_A1_JF_csp_cost_functions.R
\name{voxelize_points_mean_attributes}
\alias{voxelize_points_mean_attributes}
\title{helper function to voxelize a las element}
\usage{
voxelize_points_mean_attributes(las, res)
}
\arguments{
\item{las}{a lidR::LAS element}

\item{res}{voxel resolution in meter}
}
\value{

}
\description{
Calculate voxel mean values for all numeric attributes in the las@data table
including the XYZ-coordinates.
}
\details{
Returns a las element with XYZ-coordinates as the voxel center and
X_gr,Y_gr,Z_gr as the center of gravity (mean point coordinates) as well as
all other numeric collumns voxel mean values with their original name.
}
\note{
%% ~~further notes~~
}
\examples{

# read example data
file = system.file("extdata", "pine_plot.laz", package="TreeLS")
tls = lidR::readTLSLAS(file)
tls |> voxelize_points_mean_attributes(1) |> lidR::plot(color = 'X_gr')

}
\seealso{
\code{\link{voxelize_points}}
}
\author{
Dr. Julian Frey <julian.frey@iww.uni-freiburg.de>
}
