% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_unlist}
\alias{fast_unlist}
\title{helper function to unlist IDs generated by dbscan::frNN}
\usage{
fast_unlist(list, l)
}
\arguments{
\item{list}{a list element created by dbscan::frNN}

\item{l}{the expected length of the result
    @author Julian Frey <julian.frey@iww.uni-freiburg.de>
@export fast_unlist}
}
\description{
creates a vector of indices from a nested list created by dbscan::frNN
}
