% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farthest_point_sampling.R
\name{p_mat_dist}
\alias{p_mat_dist}
\title{Point distance function}
\usage{
p_mat_dist(mat, p)
}
\arguments{
\item{mat}{matrix with points as rows}

\item{p}{point to calculate distances}
}
\value{
the distances between every row of mat and p
}
\description{
calculates euclidean distances for n dimensions between a matrix of points and a single point
}
\examples{
p_mat_dist(as.matrix(cbind(runif(100),runif(100))), c(3,4))
}
